requirements.txt
beautifulsoup4==4.3.2
django-ace==1.0.2
django-appconf==1.0.1
django-ckeditor==4.4.7
django-classy-tags==0.6.1
django-debug-toolbar==1.2.2
django-extensions==1.5.0
django-filer==0.9.9
django-flatblocks==0.9
django-jsonfield==0.9.13
django-mptt==0.6.1
django-polymorphic==0.6.1
django-robots==1.0
django-suit==0.2.12
django-xmlrpc==0.1.5
Django==1.7.4
easy-thumbnails==2.2
ecdsa==0.13
Fabric==1.10.1
futures==2.2.0
gunicorn==19.2.1
html5lib==1.0b3
jsonfield==1.0.3
paramiko==1.15.2
Pillow==2.7.0
pip-tools==0.3.5
psycopg2==2.6
pycrypto==2.6.1
pyparsing==2.0.3
pytz==2014.10
six==1.9.0
sqlparse==0.1.14
Unidecode==0.04.17
